refactor: remove unused service providers and `AppPreferences`
This commit removes the following:

- `ServiceProviderBindsModule` and its associated `bind...ServiceProvider` methods from `apiServiceModule.kt` as they were unused.
- `bindAppPreferences` from `localDataSourceModule.kt` as `AppPreferences` and `AppDataStorePreferences` are no longer needed.

Additionally, `AuthenticationRemoteDataSourceImpl` now uses constructor injection (`@Inject constructor`).
The file `serviceModule.kt` has been renamed to `apiServiceModule.kt` for clarity.
